pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==3.2.2",
    "fastapi>=0.128.0",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
